Rust and WebGL: GUI
Introduce a window with a title bar to draw some text, making sure to
wrap text content properly to avoid drawing it outside of the window.
Also introduce a raw image class, and use it when creating a texture.

A method to draw primitives which takes care of the bindings has been
implemented for the GUI pipeline. The same idea could be applied to
the other pipelines as well.