Rust and WebGL: Refactor GUI
Avoid using Rc<RefCell> for the GUI object as it is not needed anymore.
Set some Context functions mutable so that we can update the GUI state.
Introduce an enum to reflect the fact that we can use either a
Renderbuffer or a Texture as a color attachment for a framebuffer.
Use an enum to create a variant object to be used as window element.
Get rid of the wrapper structure and just use WebGlTexture directly.
Do not use constant values for things that can be configured at runtime.